backend/requirements.txt
django>=4.2
gunicorn
djangorestframework
djoser
python-dotenv
drf-extra-fields
psycopg2-binary
Pillow
